fix compiler warning

+++ lib/url.c
@@ -18,7 +18,7 @@
  * This software is distributed on an "AS IS" basis, WITHOUT WARRANTY OF ANY
  * KIND, either express or implied.
  *
- * $Id: url.c,v 1.755 2008-10-10 15:54:07 patrickm Exp $
+ * $Id: url.c,v 1.756 2008-10-14 02:35:40 yangtse Exp $
  ***************************************************************************/
 
 /* -- WIN32 approved -- */
@@ -3176,13 +3176,13 @@ static CURLcode ParseURLAndFillConnection(struct SessionHandle *data,
     char *percent = strstr (conn->host.name, "%25");
     if (percent) {
       char *endp;
-      unsigned int scope = strtoul (percent + 3, &endp, 10);
+      unsigned long scope = strtoul (percent + 3, &endp, 10);
       if (*endp == ']') {
         /* The address scope was well formed.  Knock it out of the hostname.  */
         memmove(percent, endp, strlen(endp)+1);
         if (!data->state.this_is_a_follow)
           /* Don't honour a scope given in a Location: header */
-          conn->scope = scope;
+          conn->scope = (unsigned int)scope;
       } else
         infof(data, "Invalid IPv6 address format\n");
     }
